+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - Unreleased
 ### Added
 - Status bar into the MainWindow [#76](https://github.com/SergeyGadzhilov/SKVM/issues/76)
+- Notifications system [#64](https://github.com/SergeyGadzhilov/SKVM/issues/64)
+- Notifications about new version [#61](https://github.com/SergeyGadzhilov/SKVM/issues/61)
 
 ## [0.0.3] - (https://github.com/SergeyGadzhilov/SKVM/releases/tag/0.0.3) - 2024-01-29
 ### Added